Move function around, small bugfix
